Fixed version
Release 1.0c
Most fatal and some warning messages were never executed because of the
way the rules were ordered in the Schematron files. This was caused
because there were multiple rule elements with the same context match.
CreditNote validation was off due to its use of the invoice namespace
instead of the creditnote namespace

Changes:
- Moved the fatal messages to a separate pattern element
- Combined rules if they had the same context match
- Fixed a few context matches that now fired unnecessarily
- Removed unncessary test sets
- Changed file names to include the 'c' revision
- CreditNote validation files were using the namespaces of an invoice