Rewrote the rule for BR-NL-1
As the rule was, it resulted in an XLST error on a document with multiple AccountingSupplierParty/PartyLegalEntity fields.

Such documents should result in either a warning UBL-SR-09 or UBL-SR-11, not a total failure.

The fix is slightly more extensive, since we do still need to check whether at least one value is either '0106' or '0190'.