Change currencyID checks to work around problem with some validators
See issue #37 for the initial discussion.

I think the issue as it was reported is really a bug in those validator implementations, but the original implementation did have another problem; it gave errors on any currencyID attribute that did not match the document-level currencyID. Some of this was hidden by a different rule matching first, but it could happen with extensions that have a currencyID value too.

In order to both work around the validators that can't handle attribute-selection in rule context, and solve that issue, we have now specifically stated for which nodes the currency should be checked.

However, the original code had other issues